Version 0.0.12 -> 0.0.13. Various IE and Opera fixes.
Change `application/json` to `text/html` in servlet response for wider browser compatibility.
Set `encoding` attribute for IE (as it does not allow setting/changing `enctype` attribute on the fly).
Hook `onchange` event on input @type=file itself (IE does not like setting it on a root div of a widget).